Moving files out of utils
