+++ lib/mumukit/named_tempfile.rb
