+++ lib/mumukit/string.rb
